codex: Engine refactor 3, Migrate built-ins to modules
